First version of Basic Blockchain Concept Visualization for IoV
This is a sample visualization of Blockchain concept for Internet of Vehicles.

I have been looking for simple Blockchain visualization for my Msc research paper work and haven't found any simple example and thought to make one.

This demonstrate how a Blockchain based Traffic Management System will look like. In the visualization, a car (Car 2) is passing data to near by car (Car 1) and Road Side Units (RSUs) of Traffic Control System. Every block if data is hashed with previous block data o keep the basic concept of blockchain. Serial number of the block is hased with SHA1 algorithm and then again the whole set of data is encrypted in Base64 algorithm.

This visualization concept can be used for any kind of examples.